refactor: extract EP wave end trace
